extract distance mechanism to a new class
